feat: implement reset password and minor refactors
