+++ src/guards/admin.guard.ts
@@ -1,6 +1,6 @@
 import { UserRole } from '@prisma/client';
 import { CanActivate, ExecutionContext, Injectable } from '@nestjs/common';
-import { verifyJWT } from '../utils';
+import { verifyJWT } from '../others/utils';
 
 @Injectable()
 export class AdminAuthGuard implements CanActivate {
